requirements.txt
django==3.2.15
django-debug-toolbar==3.2.1
Pillow==8.2.0
environs[django]==9.3.2
django-phonenumber-field==8.1.0
phonenumbers==9.0.4
djangorestframework==3.16.0
geopy==2.4.1
requests==2.32.3
rollbar==1.3.0